Added Swift Target.
Since NSURL and URL are *NOT* the same type, compiler does magic and makes the bridge.
However for extensions defined like this compiler will not help.

The solution is to add an swift extension that calls the ObjC code.
Conclusion: Two targets are needed. One for ObjC and other for Swift.